pyproject.toml
[project]

name = "trail-betting"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "SEENOVATE", email = "contact@seenovate.com"}]

dependencies = [
    "streamlit>=1.43.0",
    "supabase>=2.13.0",
    "bcrypt>=4.3.0",
    "python-dotenv>=1.1.0",
    "pandas>=2.2.3",
    "streamlit-extras>=0.7.8",
]
